refactor: terminalOptions menu items
