+++ src/hooks/menuItemsHooks/index.ts
@@ -1,3 +1,4 @@
 export {useHelpMenuItems} from './useHelpMenuItems';
 export {useNewTerminalMenuItems} from './useNewTerminalMenuItems';
 export {useRefDropdownMenuItems} from './useRefDropdownMenuItems';
+export {useTerminalOptionsMenuItems} from './useTerminalOptionsMenuItems';
